pyproject.toml
[tool.poetry]
name = "hello-circleci"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.7.3"
numpy = "1.16.4"
pyYAML = "*"
dacite = "1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.0.1"